step 1 : 구간추가로직 리팩토링 및 도메인 테스트 작성 완료

+++ src/main/java/nextstep/subway/domain/Section.java
@@ -28,11 +28,10 @@ public class Section {
     @JoinColumn(name = "down_station_id")
     private Station downStation;
 
-    private int distance;
-
+    private Long distance;
 
     @Builder
-    public Section(Line line, Station upStation, Station downStation, int distance) {
+    public Section( Line line, Station upStation, Station downStation, Long distance) {
         this.line = Objects.requireNonNull(line);
         this.upStation = Objects.requireNonNull(upStation);
         this.downStation = Objects.requireNonNull(downStation);
