+++ src/main/java/nextstep/subway/dto/SectionRequest.java
@@ -10,6 +10,6 @@ public class SectionRequest {
 
     private Long upStationId;
     private Long downStationId;
-    private int distance;
+    private Long distance;
 
 }
